Add tests where scripts being used to decide about links, directories and deletes return no contents. Fix bugs with same.
Allow whitespace between numerical comparison operators and the number. 
I added this to try to fix another problem, which is no longer relevant,
but it seems like a reasonable thing to allow anyway.

If not running as root then don't request the "secure" version of
FileUtils.rmtree in the remove_file method. The secure mechanism doesn't
seem to work reliably when not run as root (see comments added to code
for more details).

Fix bug in handling of setup commands.  They need to be run even if the
config.xml doesn't include file/link/directory/delete instructions.
This is particularly true because the server will send a config.xml
request with just depend and setup entries if it needs an original
sum/contents.  This bug was causing the client to ignore the setup
commands at that time, thus backing up the probably non-existent
original file, only to run the setup later when the full config.xml
entry for the file was sent.  Added test for same.


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@50 59d7d9da-0ff4-4beb-9ab8-e480f87378b2