Fix handling of operator comparisons in attribute filtering.  Improve test for same.
Rescue any exceptions in configfilter! where it is called by generate
file so that we can insert the name of the config.xml file that we were
handling when the exception occurred. This should help the user with
debugging. Same logic as the exception catch/re-raise that occurs when
running user scripts.


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@76 59d7d9da-0ff4-4beb-9ab8-e480f87378b2