Added authentication support.
Add support for local requests.

Add RESTful controllers for etch configs, facts and originals. Populate
remaining RESTful methods for results and clients controllers.

Modify repo_update to update repo directories atomically.  We don't want
clients getting a partial or partially updated copy of the repository

versiontype bug fixes

Remove the line that enabled protect_from_forgery in application.rb,
thus allowing the removal of the lines from various controllers that
turned around and disabled that feature.  It is incompatible with the
way the etch server works, as POST requests to the etch server are not
form data.

Remove the line that turned off sessions for robots in application.rb,
as etch does not store sessions in the database so there's no real 
penalty for generating a session for robots.  That line was copied over
initially from nVentory, which does (or at least used to) store sessions 
in the database.

Copy the convert_includes method from nVentory.

Clean up the handling of find and paginate calls in the results and
client controllers, reducing code duplication.  Add support for XML
formatting of results in same.

Rename method in files controller from index to create.  See comment
there for reasoning.


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@89 59d7d9da-0ff4-4beb-9ab8-e480f87378b2