Add timeout to output capturing so that ill-behaved daemons don't cause etch to hang around forever.  Add unit tests for output capturing.
Note that the output capturing timeout test is commented out as the
timeout is rather long, thus causing the test to take a long time to
run.  Not quite sure the best way to handle that.

Replace instances of assert(string.include?(substring)) with
assert_match(substring, string) so that the user gets a better error if
the assertion fails.  Verified that it does a substring match if the
first arg is a string instead of a regex (a la String#sub), as the
documentation doesn't make that clear.

Add support for local requests.

Tests of client authentication

Modify start_server method to return pid in addition to port, and modify
stop_server method to take a pid to stop.  This allows folks to run more
than one server at a time.

Modify run_etch method to run etch with a test key, otherwise etch tries
to use the system's host key, which won't be readable unless you run the
tests as root, so etch just skips over attaching an authentication
signature to requests.


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@90 59d7d9da-0ff4-4beb-9ab8-e480f87378b2