Break the tests up into a number of test methods rather than one giant method. Reduces the risk of interference between the tests and makes it easier to run specific tests for debugging purposes.
Add a test for the case where we're starting to manage a file in a
directory that doesn't yet exist, and where a test is defined. I noticed
that we have a ordering issue in that case, we try to make the backup
due to the test before creating the directory structure, and making the
backup fails because the directory structure doesn't exist yet.


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@112 59d7d9da-0ff4-4beb-9ab8-e480f87378b2