Make it easier to run etch from within the distribution directory structure.  Remove the run_etch script I added as a poorly thought-out hack to accomplish the same.
Add support for requesting specific commands.

Add featuring allowing the server to tell the client to retry certain
commands on the next request. The "need sum" mechanism is used to do the
same thing for files, but as commands have no original content saving
mechanism we have to have an explicit feature for retries. The server
needs this if it is unable to generate all of the prerequisites for the
command, usually because the command depends on a file for which the
server needs original content info.

Add support for files that depend on commands and vice-versa.

Put all data that extracted from the server's response into a
responsedata hash rather than a bunch of separate variables. Previous
process_file was passed only file configuration and process_commands
passed only command configuration, but with the new
files-that-depend-on-commands and vice-versa features both methods need
all configuration data.

Fix bug in get_orig_contents so that it return the contents for
temporarily saved originals in addition to finalized originals.


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@132 59d7d9da-0ff4-4beb-9ab8-e480f87378b2