Store history logs as individual files rather than in RCS.  RCS results in a very cumbersome UI for viewing the history logs, making it unlikely they'll be used.  Individual files allows for easy viewing and inspection using standard Unix tools like ls, diff, grep, etc.  History logs in the old RCS format are converted automatically to the new format.
Add any exception message to the message that is reported to the etch
server.  Otherwise we just report a non-zero status with no message,
which is difficult to debug.  We were printing the exception message to
stderr, but most users run etch in a cron job that sends stderr to
/dev/null.

Update the history log after performing a revert so that we capture that
activity.  Perform another bit of code cleanup in the revert action.

Add nonrecursive_copy and nonrecursive_copy_and_rename methods to
complement the recursive versions we had already.


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@190 59d7d9da-0ff4-4beb-9ab8-e480f87378b2