Move the two dashboard charts to partials so that they can be accessed by themselves (for possible inclusion in an Architecture dashboard).
Move the chart generation logic into a separate method in the dashboard
controller so that we don't compute the client status counts when
generating the client count chart.  We need the client status counts
when generating the dashboard itself, and when generating the client
status chart.  It's bad enough computing them twice, computing them a
third time unnecessarily bugged me (and the dashboard loads rather
slowly in production, so this should help a bit).

Clean up the loop when computing the month-by-month client counts, the
old code worked fine but was ugly.

Simplify the formatting of the months on the month-by-month client count
chart, the fancier formatting I had cut-n-pasted from somewhere didn't
look very good.

Add a style to the chart titles to make them stand out more.

Wrap the charts in <div> and apply formatting in the stylesheet to
center them.  Makes the dashboard look a bit better, although there's
still some non-optimal formatting going on.  (The status chart moves
off-center sooner than it seems like it should when narrowing the
browser window.)


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@200 59d7d9da-0ff4-4beb-9ab8-e480f87378b2