https://sourceforge.net/apps/trac/etch/ticket/8 Add support for an /etc/etch.conf config file to the client. This allows users to set the etch server, the directory to use for local configuration (if desired instead of using a server), and the SSH key permanently. These were previously only configurable on the command line. The config file also allows the user to set a PATH that will be in force for any commands etch runs.  The PATH was previously hard-coded in the client.
In Etch::Client.initialize replace the :varbase option with
:file_system_root.  The :varbase option allowed the test suite to
specify an alternative to /var/etch for the client so that the test
suite doesn't have to run as root and doesn't interfere with actual etch
client usage on the system where the test suite is run. The replacement
:file_system_root option allows the test suite to define a base
directory under which the client will look for both /var/etch and
/etc/etch.conf. This allows the test suite to test various config file
settings, again without interfering with actual etch client usage on the
development system.

Change --test-base to --test-root in the etch executable to reflect the
Etch::Client option change.

In the start_server method in the test suite implement a check that the
server has started up before returning, rather than sleeping for 5
seconds and hoping that is enough time. Under heavy load it sometimes
isn't enough time, causing the test suite to fail unnecessarily. And
under light load the server starts up much quicker than 5 seconds, so it
is nice to not have to wait around unnecessarily.

Update the run_etch method in the test suite to take optional arguments
in an options hash, rather than as individual arguments. This cleans up
the method signature and makes the code that calls run_etch more
self-documenting.  Added options to allow callers to override the
--server and --key options passed to etch, and an option for specifying
the testname so that we can pass that to assert.


git-svn-id: https://etch.svn.sourceforge.net/svnroot/etch/trunk@225 59d7d9da-0ff4-4beb-9ab8-e480f87378b2